style ai/requirements.txt
flask
python-dotenv
opencv-python
numpy
pillow